+++ mvc/src/main/java/com/interface21/webmvc/servlet/view/View.java
@@ -1,9 +1,9 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.view;
+
+import java.util.Map;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
-import java.util.Map;
 
 public interface View {
     void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
